src/main/java/br/edu/ifpb/dac/ecommerce/business/service/AuthenticationServiceImpl.java
package br.edu.ifpb.dac.ecommerce.business.service;

import br.edu.ifpb.dac.ecommerce.model.entity.User;
import lombok.RequiredArgsConstructor;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class AuthenticationServiceImpl implements AuthenticationService {

    private final TokenService tokenService;

    private final UserService userService;

    private final AuthenticationManager authenticationManager;

    @Override
    public String login(String email, String password) {
        var auth = new UsernamePasswordAuthenticationToken(email, password);
        authenticationManager.authenticate(auth);

        User user = userService.findByEmail(email);
        return tokenService.generate(user);
    }

    @Override
    public User getLoggedUser() {
        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
        return (User) authentication.getPrincipal();
    }

}
